Correction validation test postman

+++ api/src/main/java/com/librairy/api/repository/CopyRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CopyRepository extends CrudRepository<Copy, Long> {
+public interface CopyRepository extends CrudRepository<Copy, Integer> {
 }
