+++ api/src/main/java/com/librairy/api/repository/LendingRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface LendingRepository extends CrudRepository<Lending, Long> {
+public interface LendingRepository extends CrudRepository<Lending, Integer> {
 }
