+++ api/src/main/java/com/librairy/api/repository/LibrairyRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface LibrairyRepository extends CrudRepository<Librairy, Long> {
+public interface LibrairyRepository extends CrudRepository<Librairy, Integer> {
 }
